requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
h5py==3.14.0
scipy
Pillow==8.4.0
matplotlib==3.5.2
    # via -r requirements.in
numpy==2.2.6
    # via
    #   -r requirements.in
    #   h5py
    #   opencv-python
    #   spectral
opencv-python==4.11.0.86
    # via -r requirements.in
pyqt6==6.9.1
    # via -r requirements.in
pyqt6-qt6==6.9.1
    # via pyqt6
pyqt6-sip==13.10.2
    # via
    #   -r requirements.in
    #   pyqt6
pyserial==3.5
    # via -r requirements.in
spectral==0.24
    # via -r requirements.in
